requirements.txt
jupyter==1.0.0
langchain==0.3.7
langgraph==0.2.45
python-dotenv==1.0.0
pydantic==2.9.2
langchain-gigachat==0.3.12
pyyaml==6.0
pytest==7.4.0
